+++ whatsnew.md
@@ -1,7 +1,9 @@
 #v1.3
 ----------------------
 
+* Method to update the layer mask of all existing meshes to make them invisible to the GUI camera
 * The text doesn't look crisp anymore !! YEAHHH !
 * Added a fade out animation for a gui object
 * Added a fade in animation for a gui object
 * Tested with Babylon 2.2 
+